+++ packages/components/src/components/Address/Address.stories.tsx
@@ -2,19 +2,21 @@ import React from 'react';
 import { Address } from '.';
 
 import { useWallet, Provider, NETWORKS } from '@web3-ui/hooks';
-import { Button } from '@chakra-ui/react';
+import { Button, VStack } from '@chakra-ui/react';
 
 export default {
   title: 'Components/Address',
   component: Address,
 };
 
-export const Default = () => <Address value='testaddress.eth' />;
+export const Default = () => <Address value="testaddress.eth" />;
 
-export const DefaultShortenedWithENS = () => <Address shortened value='testaddress.eth' />;
+export const DefaultShortenedWithENS = () => (
+  <Address shortened value="testaddress.eth" />
+);
 
 export const DefaultShortenedWithHexAddress = () => (
-  <Address shortened value='0x7Be8076f4EA4A4AD08075C2508e481d6C946D12b' />
+  <Address shortened value="0x7Be8076f4EA4A4AD08075C2508e481d6C946D12b" />
 );
 
 type AddressProps = {
@@ -25,13 +27,18 @@ const AddressUsingProvider = (props: AddressProps) => {
   const { connected, connectWallet, connection } = useWallet();
 
   return (
-    <>
+    <VStack>
       <Address
-        value={connected ? connection.ens || connection.userAddress || '' : 'Not connected'}
+        copiable
+        value={
+          connected
+            ? connection.ens || connection.userAddress || ''
+            : 'Not connected'
+        }
         shortened={props.shortened}
       />
       <Button onClick={connectWallet}>Connect wallet</Button>
-    </>
+    </VStack>
   );
 };
 
@@ -46,3 +53,5 @@ export const WithWalletShortened = () => (
     <AddressUsingProvider shortened />
   </Provider>
 );
+
+export const CanBeCopied = () => <Address value="0x00000000000000" copiable />;