+++ packages/components/src/components/NFTGallery/NFTGallery.stories.tsx
@@ -5,17 +5,16 @@ import { NFTGallery } from '.';
 export default {
   title: 'Components/NFTGallery',
   component: NFTGallery,
-  parameters: {
-    // TODO: Fix window.ethereum is undefined breaking chromatic
-    chromatic: { disableSnapshot: true },
-  },
 };
 
 export const nftsOwnedByAnAccount = () => (
-  <NFTGallery address='0x1A16c87927570239FECD343ad2654fD81682725e' />
+  <NFTGallery
+    address="0x1A16c87927570239FECD343ad2654fD81682725e"
+    gridWidth={2}
+  />
 );
 
-export const nftsOwnedByAnENS = () => {
+const nftsOwnedByAnENSStory = () => {
   const [provider, setProvider] = useState<ethers.providers.Web3Provider>();
 
   useEffect(() => {
@@ -27,7 +26,14 @@ export const nftsOwnedByAnENS = () => {
     return <>Loading...</>;
   }
 
-  return <NFTGallery address='dhaiwat.eth' web3Provider={provider} />;
+  return <NFTGallery address="dhaiwat.eth" web3Provider={provider} />;
 };
 
-export const WithAnError = () => <NFTGallery address='bad_address' />;
+export const nftsOwnedByAnENS = nftsOwnedByAnENSStory.bind({});
+// @ts-expect-error attaching parameters is fine when binding stories
+nftsOwnedByAnENS.parameters = {
+  // disables Chromatic's snapshotting on a story level
+  chromatic: { disableSnapshot: true },
+};
+
+export const WithAnError = () => <NFTGallery address="bad_address" />;