+++ packages/hooks/src/hooks/index.ts
@@ -1,2 +1,7 @@
 export { useWallet } from './useWallet';
 export { useContract } from './useContract';
+export { useTransaction } from './useTransaction';
+export { useTokenBalance } from './useTokenBalance';
+export { useReadOnlyProvider } from './useReadOnlyProvider';
+export { useReadOnlyContract } from './useReadOnlyContract';
+export { usePoller } from './usePoller';