+++ packages/hooks/src/hooks/useWallet.ts
@@ -22,6 +22,8 @@ export function useWallet() {
     connected,
     provider,
     correctNetwork,
+    network,
+    readOnlyProvider,
   } = context;
 
   React.useEffect(() => {
@@ -31,6 +33,45 @@ export function useWallet() {
       });
     }
   }, [userAddress, provider]);
+
+  const switchToCorrectNetwork = async () => {
+    if (window.ethereum) {
+      try {
+        console.log('chainId', { chainId });
+        console.log('requiredNetwork', { network });
+        // check if the chain to connect to is installed
+        await window.ethereum.request({
+          method: 'wallet_switchEthereumChain',
+          params: [{ chainId: `0x${network}` }], // chainId must be in hexadecimal numbers
+        });
+      } catch (error) {
+        console.error(error);
+        // This error code indicates that the chain has not been added to MetaMask.
+        // If it is not, then install it into the user's MetaMask
+        // @ts-expect-error unknown error type
+        if (error.code === 4902) {
+          try {
+            await window.ethereum.request({
+              method: 'wallet_addEthereumChain',
+              params: [
+                {
+                  chainId: '0x61',
+                  rpcUrl: 'https://data-seed-prebsc-1-s1.binance.org:8545/',
+                },
+              ],
+            });
+          } catch (addError) {
+            console.error(addError);
+          }
+        }
+      }
+    } else {
+      // if window.ethereum is undefined then MetaMask is not installed
+      alert(
+        'Switching networks automatically is only supported in MetaMask. Please consider installing it: https://metamask.io/download.html'
+      );
+    }
+  };
 
   return {
     connectWallet,
@@ -44,5 +85,7 @@ export function useWallet() {
     connected,
     provider,
     correctNetwork,
+    switchToCorrectNetwork,
+    readOnlyProvider,
   };
 }