+++ packages/hooks/src/stories/ConnectWallet.stories.tsx
@@ -10,11 +10,18 @@ export default {
 const portisDappId = '512893f6-6436-44c3-b0dc-6caccab984bb';
 
 const DefaultUsingProvider = () => {
-  const { connection, connectWallet, disconnectWallet, connected, correctNetwork } = useWallet();
+  const {
+    connection,
+    connectWallet,
+    disconnectWallet,
+    connected,
+    correctNetwork,
+    switchToCorrectNetwork,
+  } = useWallet();
 
   useEffect(() => {
     if (!correctNetwork) {
-      alert('Please connect to the correct network');
+      alert('Please connect to Rinkeby.');
     }
   }, [correctNetwork]);
 
@@ -24,6 +31,11 @@ const DefaultUsingProvider = () => {
         <Button onClick={disconnectWallet}>Disconnect wallet</Button>
         <p>{connection.ens || connection.userAddress}</p>
         <p>Connected to the correct network: {correctNetwork ? 'Yes' : 'no'}</p>
+        {!correctNetwork && (
+          <Button colorScheme="teal" mt={2} onClick={switchToCorrectNetwork}>
+            Switch to correct network
+          </Button>
+        )}
       </div>
     );
   }