+++ packages/hooks/src/stories/UseTokenBalance.stories.tsx
@@ -1,12 +1,13 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import { Button } from '@chakra-ui/react';
+import { Button, VStack, Text } from '@chakra-ui/react';
 import { NETWORKS, Provider, useTokenBalance, useWallet } from '..';
 
 const GTC_ADDRESS = '0xde30da39c46104798bb5aa3fe8b9e0e1f348163f';
 
 const Default = () => {
-  const { connection, connectWallet, disconnectWallet, connected } = useWallet();
+  const { connection, connectWallet, disconnectWallet, connected } =
+    useWallet();
   const { balance, formattedBalance, loading, error } = useTokenBalance({
     tokenAddress: GTC_ADDRESS,
     accountAddress: connection.userAddress!, // you can test using 0x0ED6Cec17F860fb54E21D154b49DAEFd9Ca04106
@@ -14,23 +15,23 @@ const Default = () => {
 
   if (connected) {
     return (
-      <div>
+      <VStack>
         <Button onClick={disconnectWallet}>Disconnect wallet</Button>
-        <p>{connection.ens || connection.userAddress}</p>
+        <Text>{connection.ens || connection.userAddress}</Text>
         {error ? (
-          <p>Error occured while trying to fetch balance.</p>
+          <Text>Error occured while trying to fetch balance.</Text>
         ) : loading ? (
-          <p>Loading...</p>
+          <Text>Loading...</Text>
         ) : (
-          <p>
+          <Text>
             GTC balance: {balance} wei, {formattedBalance} GTC
-          </p>
+          </Text>
         )}
-      </div>
+      </VStack>
     );
   }
 
-  return <button onClick={connectWallet}>Connect Wallet</button>;
+  return <Button onClick={connectWallet}>Connect Wallet</Button>;
 };
 
 storiesOf('Hooks/useTokenBalance', module).add('Default', () => (