+++ packages/components/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @web3-ui/components
 
+## 0.5.0
+
+### Minor Changes
+
+- [#90](https://github.com/Developer-DAO/web3-ui/pull/90) [`16b8481`](https://github.com/Developer-DAO/web3-ui/commit/16b8481995f3f934f94305fbae5f3a6370e4fb20) Thanks [@swetshaw](https://github.com/swetshaw)! - Added a EtherInput component that supports input in wei or ether and automatically converts it to wei for you!
+
+## 0.4.1
+
+### Patch Changes
+
+- [#170](https://github.com/Developer-DAO/web3-ui/pull/170) [`53da320`](https://github.com/Developer-DAO/web3-ui/commit/53da320bdaeb28ac4b82152b7c289f7aa9d7b695) Thanks [@Dhaiwat10](https://github.com/Dhaiwat10)! - Add README for the components package
+
+## 0.4.0
+
+### Minor Changes
+
+- [#150](https://github.com/Developer-DAO/web3-ui/pull/150) [`f0c96df`](https://github.com/Developer-DAO/web3-ui/commit/f0c96df7468fbe0993b8e90979407c27ba7a22bd) Thanks [@Ibby-devv](https://github.com/Ibby-devv)! - Accept any ChakraProvider-compatible props in our Provider
+
+* [#163](https://github.com/Developer-DAO/web3-ui/pull/163) [`01ce580`](https://github.com/Developer-DAO/web3-ui/commit/01ce5809debb2284545620861d11893e4f9675f0) Thanks [@Dhaiwat10](https://github.com/Dhaiwat10)! - Add missing exports
+
+- [#164](https://github.com/Developer-DAO/web3-ui/pull/164) [`184ec43`](https://github.com/Developer-DAO/web3-ui/commit/184ec43e0542a0057f7847a9c696eb0479f96438) Thanks [@hone1er](https://github.com/hone1er)! - Bug fix. Added null check to address component for value prop
+
+## 0.3.1
+
+### Patch Changes
+
+- [#149](https://github.com/Developer-DAO/web3-ui/pull/149) [`4c2056d`](https://github.com/Developer-DAO/web3-ui/commit/4c2056d9c8b2c1eebdd7135a5581447941a65ad8) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Remove the `cross-fetch` polyfill
+
+## 0.3.0
+
+### Minor Changes
+
+- [#105](https://github.com/Developer-DAO/web3-ui/pull/105) [`6064be0`](https://github.com/Developer-DAO/web3-ui/commit/6064be033150ccd4df52c913a3d55ce2c63c262b) Thanks [@maximebonhomme](https://github.com/maximebonhomme)! - Address component now accept copiable prop to allow users to copy the address value into their clipboard
+
+* [#130](https://github.com/Developer-DAO/web3-ui/pull/130) [`825561b`](https://github.com/Developer-DAO/web3-ui/commit/825561bd2ea2abc541fe7bde59446ddea3aae49d) Thanks [@etr2460](https://github.com/etr2460)! - Fix NFT sizes
+
+- [#121](https://github.com/Developer-DAO/web3-ui/pull/121) [`aa9b2f6`](https://github.com/Developer-DAO/web3-ui/commit/aa9b2f6cccdcda1d443bacb306d0a83398a19368) Thanks [@hone1er](https://github.com/hone1er)! - Added a TokenGate component that restricts access to child components unless erc-20/erc-721 token quantity requirements are met.
+
+### Patch Changes
+
+- [#132](https://github.com/Developer-DAO/web3-ui/pull/132) [`2b7b5c7`](https://github.com/Developer-DAO/web3-ui/commit/2b7b5c7295e5fb389740a9f4fab4e3fcea7e5e7d) Thanks [@maximebonhomme](https://github.com/maximebonhomme)! - Address component now uses Text component instead of unnecesary Input
+
+## 0.2.2
+
+### Patch Changes
+
+- [#116](https://github.com/Developer-DAO/web3-ui/pull/116) [`c843555`](https://github.com/Developer-DAO/web3-ui/commit/c843555369f56c01653f0486e54a31a382353ed4) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Add babel-runtime to deps so we can have regenerator runtime at runtime
+
 ## 0.2.1
 
 ### Patch Changes