+++ packages/core/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @web3-ui/core
 
+## 0.4.1
+
+### Patch Changes
+
+- [#217](https://github.com/Developer-DAO/web3-ui/pull/217) [`0c5209b`](https://github.com/Developer-DAO/web3-ui/commit/0c5209b987665a3875b297733e6dde8205cd3965) Thanks [@with-heart](https://github.com/with-heart)! - Added `ERC20ABI` as an export of `@web3-ui/hooks` in order to resolve an incorrect import in the `@web3-ui/core` package.
+
+- Updated dependencies [[`0c5209b`](https://github.com/Developer-DAO/web3-ui/commit/0c5209b987665a3875b297733e6dde8205cd3965), [`16b8481`](https://github.com/Developer-DAO/web3-ui/commit/16b8481995f3f934f94305fbae5f3a6370e4fb20), [`ece14b2`](https://github.com/Developer-DAO/web3-ui/commit/ece14b2ea025772eef0251432dc43e41e3e219c0), [`b66e93a`](https://github.com/Developer-DAO/web3-ui/commit/b66e93a97f9c260b903ba0545fd368ec4842f349)]:
+  - @web3-ui/hooks@0.9.0
+  - @web3-ui/components@0.5.0
+
+## 0.4.0
+
+### Minor Changes
+
+- [#223](https://github.com/Developer-DAO/web3-ui/pull/223) [`d52e750`](https://github.com/Developer-DAO/web3-ui/commit/d52e75078af16aa851c834624e5130a741d66556) Thanks [@Dhaiwat10](https://github.com/Dhaiwat10)! - The TokenGate component in the core package now supports ERC721 tokens. The hooks package now also exports the ERC721 ABI
+
+### Patch Changes
+
+- Updated dependencies [[`d52e750`](https://github.com/Developer-DAO/web3-ui/commit/d52e75078af16aa851c834624e5130a741d66556)]:
+  - @web3-ui/hooks@0.8.1
+
+## 0.3.0
+
+### Minor Changes
+
+- [#214](https://github.com/Developer-DAO/web3-ui/pull/214) [`e204d49`](https://github.com/Developer-DAO/web3-ui/commit/e204d49bd8ff8c29015b63d33df44fab0b302c84) Thanks [@Dhaiwat10](https://github.com/Dhaiwat10)! - The core package now has a TokenGate that lets you conditionally render some content depending on whether a user owns some amount of a token or not.
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [[`000c2f8`](https://github.com/Developer-DAO/web3-ui/commit/000c2f8a8f15e963d090f305f7d6e6073ec41e4c), [`674078c`](https://github.com/Developer-DAO/web3-ui/commit/674078c28ea2b32229b119dd2f01bfa9a6b9e8af)]:
+  - @web3-ui/hooks@0.8.0
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [[`2b39709`](https://github.com/Developer-DAO/web3-ui/commit/2b3970987eff7c9aeabbce027477079a43a61220)]:
+  - @web3-ui/hooks@0.7.0
+
+## 0.2.0
+
+### Minor Changes
+
+- [#166](https://github.com/Developer-DAO/web3-ui/pull/166) [`1a33c80`](https://github.com/Developer-DAO/web3-ui/commit/1a33c8097a2608e029336301a0082e1d240b65df) Thanks [@Dhaiwat10](https://github.com/Dhaiwat10)! - Export all the hooks from the core package too
+
+### Patch Changes
+
+- Updated dependencies [[`f0c96df`](https://github.com/Developer-DAO/web3-ui/commit/f0c96df7468fbe0993b8e90979407c27ba7a22bd), [`01ce580`](https://github.com/Developer-DAO/web3-ui/commit/01ce5809debb2284545620861d11893e4f9675f0), [`184ec43`](https://github.com/Developer-DAO/web3-ui/commit/184ec43e0542a0057f7847a9c696eb0479f96438)]:
+  - @web3-ui/components@0.4.0
+  - @web3-ui/hooks@0.6.0
+
+## 0.1.4
+
+### Patch Changes
+
+- [#149](https://github.com/Developer-DAO/web3-ui/pull/149) [`4c2056d`](https://github.com/Developer-DAO/web3-ui/commit/4c2056d9c8b2c1eebdd7135a5581447941a65ad8) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Remove the `cross-fetch` polyfill
+
+- Updated dependencies [[`4c2056d`](https://github.com/Developer-DAO/web3-ui/commit/4c2056d9c8b2c1eebdd7135a5581447941a65ad8), [`169021d`](https://github.com/Developer-DAO/web3-ui/commit/169021df19ebc7819f68d74213dc74380d8a71d6), [`e611ef9`](https://github.com/Developer-DAO/web3-ui/commit/e611ef9860ac0179e3c3dcfc64f05aac3b88baa7)]:
+  - @web3-ui/components@0.3.1
+  - @web3-ui/hooks@0.5.1
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [[`6064be0`](https://github.com/Developer-DAO/web3-ui/commit/6064be033150ccd4df52c913a3d55ce2c63c262b), [`825561b`](https://github.com/Developer-DAO/web3-ui/commit/825561bd2ea2abc541fe7bde59446ddea3aae49d), [`da536ed`](https://github.com/Developer-DAO/web3-ui/commit/da536ed8df37c768764ca616a9bb83fc91a01d5e), [`3f0bc64`](https://github.com/Developer-DAO/web3-ui/commit/3f0bc640cc4d0d4ee4e656afd16c6fc2fbfc9cbd), [`2b7b5c7`](https://github.com/Developer-DAO/web3-ui/commit/2b7b5c7295e5fb389740a9f4fab4e3fcea7e5e7d), [`aa9b2f6`](https://github.com/Developer-DAO/web3-ui/commit/aa9b2f6cccdcda1d443bacb306d0a83398a19368)]:
+  - @web3-ui/components@0.3.0
+  - @web3-ui/hooks@0.5.0
+
+## 0.1.2
+
+### Patch Changes
+
+- [#116](https://github.com/Developer-DAO/web3-ui/pull/116) [`c843555`](https://github.com/Developer-DAO/web3-ui/commit/c843555369f56c01653f0486e54a31a382353ed4) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Add babel-runtime to deps so we can have regenerator runtime at runtime
+
+- Updated dependencies [[`c843555`](https://github.com/Developer-DAO/web3-ui/commit/c843555369f56c01653f0486e54a31a382353ed4)]:
+  - @web3-ui/components@0.2.2
+  - @web3-ui/hooks@0.4.1
+
 ## 0.1.1
 
 ### Patch Changes